+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -22,19 +22,26 @@ class BootStrap implements LogSupport {
     def logLevelService,
         configService,
         clusterService,
+        metricsService,
         prefService
 
     def init = {servletContext ->
         logStartupMsg()
-        ensureRequiredConfigsCreated()
-        ensureRequiredPrefsCreated()
 
+        parallelInit([configService])
+        ensureRequiredConfigsCreated()
         ensureExpectedServerTimeZone()
 
-        def services = Utils.xhServices.findAll {it.class.canonicalName.startsWith('io.xh.hoist')}
+        // Ordered, early initialization of core service used by other services.
         parallelInit([logLevelService])
         parallelInit([clusterService])
+        parallelInit([metricsService])
+
+        // All other services in parallel
+        def services = Utils.xhServices.findAll {it.class.canonicalName.startsWith('io.xh.hoist')}
         parallelInit(services)
+
+        ensureRequiredPrefsCreated()
     }
 
     def destroy = {}
@@ -300,6 +307,17 @@ class BootStrap implements LogSupport {
                 defaultValue: 'none',
                 groupName: 'xh.io',
                 note: 'Email address to which status monitor alerts should be sent. Value "none" disables emailed alerts.'
+            ],
+            xhMetricsConfig: [
+                valueType: 'json',
+                defaultValue: [
+                    prometheusEnabled: false,
+                    otlpEnabled: false,
+                    prometheusConfig: [:],
+                    otlpConfig: [:]
+                ],
+                groupName: 'xh.io',
+                note: 'Parameters for observable metric support'
             ],
             xhWebSocketConfig: [
                 valueType: 'json',