+++ grails-app/services/io/xh/hoist/monitor/MonitorService.groovy
@@ -38,6 +38,7 @@ class MonitorService extends BaseService {
     ConfigService configService
     MonitorReportService monitorReportService
     MonitorEvalService monitorEvalService
+    MonitorMetricsService monitorMetricsService
 
     // Shared state for all servers to read - gathered by primary from all instances.
     // Map of monitor code to aggregated (cross-instance) results.
@@ -56,6 +57,16 @@ class MonitorService extends BaseService {
             delay: startupDelay,
             primaryOnly: true
         )
+    }
+
+    /**
+     * Get the current set of aggregated results for all configured monitors.
+     *
+     * Results will be sorted according to sort orders defined in the monitor configurations and
+     * will include stub entries for inactive monitors.
+     */
+    AggregateMonitorResult getResult(String code) {
+        _results.get()?[code]
     }
 
     /**
@@ -100,6 +111,9 @@ class MonitorService extends BaseService {
             [code, newResults(checks, prevResults?[code])]
         }
         _results.set(newResults)
+
+        // Publish per-instance metrics to Micrometer
+        monitorMetricsService.noteResultsUpdated(newResults.values())
 
         // Report the canonical results from public getter
         monitorReportService.noteResultsUpdated(results)